requirements.txt
flask==3.0.2
minio==7.2.5
python-dotenv==1.0.1
flask-wtf==1.2.1 